requirements.txt
pandas 
requests
psycopg2
psycopg2-binary
sqlalchemy
python-dotenv